[autotest] Allow running deployment_test on Repair Failed DUTs.
The current deployment_test script requires that a DUT be locked
and in the 'Ready' state.  However, an identified use case for the
deployment_test script is to use it during DUT repair after
repairing a servo; in this case the DUT state will more likely be
'Repair Failed'.

This changes the script support the additional use case.

BUG=None
TEST=Run against a failed device in the test lab

Change-Id: Ic7a2c664402524e12ece17f344e8d6923f17d232
Reviewed-on: https://chromium-review.googlesource.com/297483
Tested-by: Richard Barnette <jrbarnette@chromium.org>
Reviewed-by: Dan Shi <dshi@chromium.org>