Allow passing additional arguments to telemetry scripts.
I am moving AFDO profile collection to use telemetry benchmarks.
For this, I need to reduce number of iterations in the page cycler
benchmarks so that collecting the profile information can be done
in less time. So, I need to pass the '--pageset-repeat' option to
the telemetry run_benchmark script.

BUG=None
TEST=Tested by executing with several telemetry benchmarks.

Change-Id: Ib69e25a0c6ea30d59b4082c8d9a3d3b3675045aa
Reviewed-on: https://chromium-review.googlesource.com/298060
Commit-Ready: Luis Lozano <llozano@chromium.org>
Tested-by: Luis Lozano <llozano@chromium.org>
Reviewed-by: Achuith Bhandarkar <achuith@chromium.org>
Reviewed-by: Caroline Tice <cmtice@chromium.org>