AFDO: Use Telemetry and Superhero to collect AFDO profile.
Until now we have been avoiding using Telemetry directly to
run the benchmarks used in the profile collection. Telemetry
is now much more stable on ChromeOS so we can give it a try.
I also updated the list of benchmars in the profile to base
it on the Superhero benchmarks. This brings more "diversity"
to the profile set.
Also, added logic to retry benchmarks in case of failure.

With these chanages we get upto 19% improvement on some
benchmarks.

BUG=None
TEST=Collected profile several times. Tested retry logic.

Change-Id: I138cadf904e40fcbc8b8d6936c65d887dd0bca5e
Reviewed-on: https://chromium-review.googlesource.com/298062
Commit-Ready: Luis Lozano <llozano@chromium.org>
Tested-by: Luis Lozano <llozano@chromium.org>
Reviewed-by: Luis Lozano <llozano@chromium.org>