Chameleon: Fix return value of move_to_display().
Under current implementation of move_to_display(), if it returns
True, it doesn't mean that the window is successfully moved to
the indicated screen, but just means that the actions are done.
This will confuse the user since they can't detect whether it
really succeed without taking a look at the screen.

This patch fixes this problem by adding a check process which
really checks whether the result window is as expected.

BUG=chromium:530308
TEST=Ran display_Resolution.extended on nyan_blaze and it passed.
     Ran display_ResolutionList.extended on nyan_blaze with issue
     529829 unhandled, under which move_to_display() should fail,
     and move_to_display() returned False as expected.

Change-Id: I62c02d58ecac618166505df477ee66863af7760a
Reviewed-on: https://chromium-review.googlesource.com/299040
Commit-Ready: Ting-Yuan Cheng <tingyuan@chromium.org>
Tested-by: Chin-Huang Lin <chinlin@google.com>
Reviewed-by: Yuli Huang <yuli@chromium.org>