[autotest] Skip RPC retry when getting JSONRPCException.
frontend.afe.json_rpc.proxy.ValidationError is a subclass of
JSONRPCException. Getting JSONRPCException for an RPC means
that the caller was able to call a function in the remote but
got an error. Because the remote function call was done but
got an error, retrying is not neccessary. Ohterwise, we will
get the same error again.
ValidationError is included in the blacklist for retry in such
a reason. We should make the blacklist wider to include
JSONRPCException.

BUG=chromium:532746
TEST=puppylab. Try ./cli/atest label remove -m test_host15 random_label

Change-Id: Ia24df450dd6b8f98788a51ee90d262dd527d13d8
Reviewed-on: https://chromium-review.googlesource.com/300601
Commit-Ready: Mungyung Ryu <mkryu@google.com>
Tested-by: Mungyung Ryu <mkryu@google.com>
Reviewed-by: Dan Shi <dshi@chromium.org>