[autotest] Forward user when forwaring/routing RPCs.
User account information is passed to RPC server via
HTTP header. We can pass the user info to the header when
creating server.frontend.RpcClient object. We should forward
the user info when forwaring/routing RPCs not to lose the
user account info in the receiving side.

BUG=chromium:533159
TEST=puppylab. Lock a sharded host with root account and check
the host is locked by root in the master DB and the shard DB.

Change-Id: I185e49f028c8fdfdf630855412a6e370d5684491
Reviewed-on: https://chromium-review.googlesource.com/300594
Commit-Ready: Mungyung Ryu <mkryu@google.com>
Tested-by: Mungyung Ryu <mkryu@google.com>
Reviewed-by: Fang Deng <fdeng@chromium.org>
Reviewed-by: Dan Shi <dshi@chromium.org>