Detect bad TPM state so repair is triggered.
The powerwash operation now clears the TPM when possible so no
modification to the repair code is needed. The specific case targeted
here is when the TPM Storage Root Key (SRK) is restricted and the public
key cannot be read. As a result, new keys cannot be imported and tests
which rely on importing keys will fail every time.

BUG=chromium:514037
TEST=manually called verify_software() with a dummy test and checked
correct result with different TPM states

Change-Id: Ida78457ec0e581c04d7f0c4e6c82df13638fc62f
Reviewed-on: https://chromium-review.googlesource.com/302120
Commit-Ready: Darren Krahn <dkrahn@chromium.org>
Tested-by: Darren Krahn <dkrahn@chromium.org>
Reviewed-by: Darren Krahn <dkrahn@chromium.org>