[autotest] audio_widget: Lower the default scale of MIC
Previously, in the CL
b8f9f6 [autotest] Use louder test file to fix internal speaker test
we use a louder file on internal speaker testing so device
with smaller speaker volume can have their audio recorded clear enough.
However, this breaks the internal speaker test on device with louder
speaker volume because the recorded raw data will be scaled by MIC's
default gain, and that default scale (300.0) is too large such that
the scaled data overflows.
We should lower the default scale of MIC because the purpose of that
scale is just used to avoid the data being too low to be misjudged as
not meaningful when doing comparison.

BUG=chromium:534692
TEST=run audio_AudioBasicInternalSpeaker test on nyan_big(small volume)
and link(large volume), and squawks.
Change-Id: Iadad1cec991bb20e00012c5836ad4bbaf4a15a2c
Reviewed-on: https://chromium-review.googlesource.com/301490
Commit-Ready: Cheng-Yi Chiang <cychiang@chromium.org>
Tested-by: Cheng-Yi Chiang <cychiang@chromium.org>
Reviewed-by: Wai-Hong Tam <waihong@chromium.org>