[autotest] Confirm servo is properly initialized in provision_FirmwareUpdate.
If a provision job updates both CrOS and FW, the CrosHost object created in
provision special task has required_by_test set to False, thus servo is not
verified. That leads to the provision_FirmwareUpdate to fail.

This change add a check (confirm_servo) in provision_FirmwareUpdate to make
sure servo object is created with required_by_test set to True, which
guarantees the servo is initialized and verified.

BUG=chromium:520656
TEST=local run and unittest

Change-Id: Ia4b86fa36b914efa2beec768099de91e5d3dcef8
Reviewed-on: https://chromium-review.googlesource.com/301296
Commit-Ready: Dan Shi <dshi@chromium.org>
Tested-by: Dan Shi <dshi@chromium.org>
Reviewed-by: Fang Deng <fdeng@chromium.org>