autotest: WiFi_RegDomain: don't use CaptureResult.pcap_path
We want to remove the |pcap_path| field from CaptureResult, so
network_WiFi_RegDomain needs to stop using |pcap_path|.

As it turns out, the way that RegDomain was using pcap_path
was rather unnecessary. In fact, stop_capture() will copy
the file over for us automatically.

A small side-effect of this change is that the pcap for the
scan phase will no longer be in the debug/ subdirectory.
However, that's a good thing. Instead of being in debug/,
the scan phase's pcap file will be in the same directory
as the pcap files for the other phases.

BUG=chromium:535362
TEST=network_WiFi_RegDomain (bcm435x, ca.hybird, ch 52+164)

Change-Id: I5674325501e8fa39da94aa9487a371a34716c306
Reviewed-on: https://chromium-review.googlesource.com/302221
Commit-Ready: mukesh agrawal <quiche@chromium.org>
Tested-by: mukesh agrawal <quiche@chromium.org>
Reviewed-by: Paul Stewart <pstew@chromium.org>