autoupdate_EndToEnd: Remove servo support, part 1 (reprise).
[Second attempt, this time taking into account deployment to prod]

We don't expect to ever be using this feature so deprecating servo
support from the test code and making the servo_host argument optional.
Once deployed, we'll remove the servo argument from the test entirely,
as well as cleanup the control file, test_params.py and
full_release_test.py.

BUG=b:24405484
TEST=None

Change-Id: I0a937f62d017b541c38613b76d9c59abb3ef80a2
Reviewed-on: https://chromium-review.googlesource.com/302871
Reviewed-by: Simran Basi <sbasi@chromium.org>
Tested-by: Gilad Arnold <garnold@chromium.org>