[autotest] Allow provision job to pick devserver from any subnet
This change removes the limit on provision job to pick a devserver from its
own subnet if possible. Lab recently added some bandwidth between the two
test labs, so it should be OK for a dut to use devserver from another lab.
This is mainly to reduce the load of devserver.

BUG=None
TEST=None

Change-Id: I8497f57771939118a7f635a4a6a05da1cd6dded5
Reviewed-on: https://chromium-review.googlesource.com/302451
Commit-Ready: Dan Shi <dshi@chromium.org>
Tested-by: Dan Shi <dshi@chromium.org>
Reviewed-by: Fang Deng <fdeng@chromium.org>