[autotest] Add SSHHost.run_background()
A common idiom for certain commands is to say "start this command in
the background, but don't wait for completion".  This occurs most
commonly with commands such as reboot that will cause a temporary
loss of network connectivity.  Because of certain obscure sshd
behaviors, the idiom is more complicated than a mere '&' in the
command.

This change adds SSHHost.run_background() to provide a method to
implement the idiom correctly, consistently, and understandably.
This also changes all instances of the idiom to use the method.

BUG=None
TEST=test_that platform_BootPerfServer

Change-Id: I0edaf36dc10ac67541aa55cf8cc4c6542878fb79
Reviewed-on: https://chromium-review.googlesource.com/302417
Commit-Ready: Richard Barnette <jrbarnette@chromium.org>
Tested-by: Richard Barnette <jrbarnette@chromium.org>
Reviewed-by: Richard Barnette <jrbarnette@chromium.org>