[autotest] teach run_suite to dump time stats to match suite bucket
Since we switched to use suite_attr_wrapper, we don't not send stats to
the right stats buckets that are named after the actual suites name
(everything is dumped to the bucket "suite_attr_wrapper"). Fix this
issue to dump to the right suite bucket.

BUG=chromium:536110
TEST=unittest and test on local autotest with some debug logging.

Change-Id: I3d99798a9093d9a19ca3a4fc614e519286be51de
Reviewed-on: https://chromium-review.googlesource.com/302952
Commit-Ready: Shuqian Zhao <shuqianz@chromium.org>
Tested-by: Mungyung Ryu <mkryu@google.com>
Reviewed-by: Mungyung Ryu <mkryu@google.com>