[autotest] Add serial flag for adb_host support in cli/atest.
User should be able to specify serials to pass to adb_host and
the board label will be autodetected and the serial will be
added as a host attribute.  Currently only works only for 1
serial, a later cl will enable multiple serials.

Also refactored the label decorator logic into site_utils
and remote.RemoteHost.

BUG=chromium:535352
TEST=manually added hosts using the serial flag using cli/atest
and checking the host was correctly detected on a moblab.

Change-Id: I4c95377eaee4529ea99add4387be6dcf56a529fb
Reviewed-on: https://chromium-review.googlesource.com/303214
Commit-Ready: Kevin Cheng <kevcheng@chromium.org>
Tested-by: Kevin Cheng <kevcheng@chromium.org>
Reviewed-by: Kevin Cheng <kevcheng@chromium.org>