Autotest: Add common test to run stressed versions of client tests
Create stress_ClientTestReboot test that, passed a client test name
and an number of iterations, will:

for N iterations:
    - run the client test on the DUT
    - reboot the machine

The server side test will not stop if there is a failure on the
client side test.  This behavior is intentional.

Control files for touch and bluetooth sanity tests are included.

TEST=ran all three control files
BUG=None

Change-Id: I6e90bc84654c4e29a15372ebffe4e48befd9b5be
Reviewed-on: https://chromium-review.googlesource.com/303187
Commit-Ready: Katherine Threlkeld <kathrelkeld@chromium.org>
Tested-by: Katherine Threlkeld <kathrelkeld@chromium.org>
Reviewed-by: danny chan <dchan@chromium.org>