autoupdate_EndToEnd: Enhance reporting of update errors.
This utilizes previously introduced logic for adding detailed error
messages when various update events are not as expected or entirely
missing. The intent for messages is to be informative, comprehensive and
actionable.

We'll improve reporting of other test failures (provisioning, sanity
checks) in subsequent CLs.

BUG=b:24158510
TEST=None

Change-Id: I03e0b173251f462cba413e63e71ff94c6e7c67be
Reviewed-on: https://chromium-review.googlesource.com/303168
Commit-Ready: Gilad Arnold <garnold@chromium.org>
Tested-by: Gilad Arnold <garnold@chromium.org>
Reviewed-by: Simran Basi <sbasi@chromium.org>