Implement policy_ImagesBlockedForUrls AutoTest
This test verifies the behavior of Chrome OS for a range of valid
values of the ImagesBlockedForUrls user policy, using four test
unique cases: NotSet, 1Url, 2Urls, and 3Urls.

When the policy is undefined (case=NotSet), images are not blocked
on any page. When set to one or more URLs, images are blocked only
on pages with the same domain as any of the listed URLs.

BUG=chromium:533712
TEST=Run test_that <IP> policy_ImagesBlockedForUrls

Change-Id: I3d20f76f6dd117d00331bf8fdd0e3d4e8e4938ac
Reviewed-on: https://chromium-review.googlesource.com/301115
Commit-Ready: Scott Cunningham <scunningham@chromium.org>
Tested-by: Scott Cunningham <scunningham@chromium.org>
Reviewed-by: David Haddock <dhaddock@chromium.org>
Reviewed-by: Scott Cunningham <scunningham@chromium.org>