autoupdate_EndToEnd: Convert all attributes to strings unconditionally.
Most attributes are encoded as strings on the devserver event log, but
some (e.g. event_result) are integers. Due to increased failure logging
verbosity, this is biting us more often. This change enforces a string
representation as soon as events are harvested, which solves this
problem.

BUG=b:24158510
TEST=Failure message generated correctly, KeyError not raised.

Change-Id: Iead4fd87ab9e3bda9347a71ea0623fc95874d5c1
Reviewed-on: https://chromium-review.googlesource.com/304610
Commit-Ready: Gilad Arnold <garnold@chromium.org>
Tested-by: Gilad Arnold <garnold@chromium.org>
Reviewed-by: Simran Basi <sbasi@chromium.org>