Implement policy_ImagesAllowedForUrls test
This test verifies the effect of the ImagesAllowedForUrls user
policy on Chrome OS client behavior and appearance, when policy
DefaultImagesSetting=2 (block all but allowed). It exercises a
range of policy values using four unique test cases, named:
NotSet, 1Url, 2Urls, and 3Urls.

The test shall pass if the browser shows an image on a test page
in a domain that is listed in the ImagesAllowedForUrls policy.
The test shall fail if the browser blocks an image on a test page
in such a domain.

BUG=chromium:538148
TEST=Run test_that <IP> policy_ImagesAllowedForUrls

Change-Id: I561940de5a01479b12516718eedcf700329cec71
Reviewed-on: https://chromium-review.googlesource.com/303544
Commit-Ready: Scott Cunningham <scunningham@chromium.org>
Tested-by: Scott Cunningham <scunningham@chromium.org>
Reviewed-by: Katherine Threlkeld <kathrelkeld@chromium.org>