[autotest] For lab inventory, don't recommend non-inventory boards
The regular board inventory excludes boards that have no suites
pool, or that have only suites DUTs.  However, when we create a list
of DUTs for repair, we don't honor those exclusions.  The result is
that in some cases, we recommend DUTs for repair when the board
isn't listed in the inventory.

This change makes the repair recommendations be drawn only from
the boards that are part of the regular inventory.

BUG=None
TEST=unit tests, plus selected testing w/ --debug

Change-Id: I0dfbd060193afbd9dd08d646151e598a9b8c0d5f
Reviewed-on: https://chromium-review.googlesource.com/305398
Commit-Ready: Richard Barnette <jrbarnette@chromium.org>
Tested-by: Richard Barnette <jrbarnette@chromium.org>
Reviewed-by: Paul Hobbs <phobbs@google.com>