autotest (wifi): Do not try to write to /tmp on an ADBHost
Because the root filesystem is read-only and writing to /data/local/tmp
is a more appropriate location on that platform.

The file will now be placed into a subdirectory named autoserv-<random>
under the base temporary directory (be it /tmp or /data/local/tmp) for
each test run. The temporary directories all get cleaned up after the
end of the test run.

BUG=chrome-os-partner:42310
TEST=Ran network_WiFi_SimpleConnect through adb successfully.
  Also ran it on samus and confirmed it did not cause any regressions.

Change-Id: I1f7e94c3aad91c05cd958698a44c491b4ed673ee
Signed-off-by: Filipe Brandenburger <filbranden@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/306115
Reviewed-by: Christopher Wiley <wiley@chromium.org>
Reviewed-by: Bindu Mahadev <bmahadev@chromium.org>
Reviewed-by: Simran Basi <sbasi@chromium.org>