brillo_test_launcher: Install the chromite site-package if missing.
brillo_test_launcher requires the chromite site-package to be present
in order to import a number of important autotest libraries.

If it is missing, build_externals is called and will install it.

Also without a complete site-packages, logging needs to be configured
manually now.

BUG=b:25118912
TEST=rm -rf site-packages/*; ./brillo_test_launcher run.

Change-Id: Ia14d81a580d5ed2748cca4107f8751b4d2ab5656
Reviewed-on: https://chromium-review.googlesource.com/307238
Commit-Ready: Simran Basi <sbasi@chromium.org>
Tested-by: Simran Basi <sbasi@chromium.org>
Reviewed-by: Dan Shi <dshi@chromium.org>