[autotest] Change to get_managed_boards() in lab inventory.
Previously, the _LabInventory class contained a `get_board_counts()`
method that returned a list of tuples for selected boards.  That
interface was bad both for readability and reusability.

This change replaces the bad API with a method called
`get_managed_boards()` that returns just a set of the boards that
wer of interest in the old `get_board_counts()` API.  Callers of
the method and the associate unit tests have been adjusted for the
change.

BUG=None
TEST=unit tests, plus selected lab_inventory --debug invocations

Change-Id: Iafcf7fe36984fe1a806337417536d541b3d1dfcf
Reviewed-on: https://chromium-review.googlesource.com/307592
Commit-Ready: Richard Barnette <jrbarnette@chromium.org>
Tested-by: Richard Barnette <jrbarnette@chromium.org>
Reviewed-by: Richard Barnette <jrbarnette@chromium.org>