autotest(wifi): Add frequency check in SimpleConnect.
Bail out of SimpleConnect tests gracefully if it detects that the DUT
does not support the frequency band being tested.

BUG=brillo:25129221
TEST=Ran network_WiFi_SimpleConnect.wifi_check11a &
network_WiFi_SimpleConnect.wifi_check11g on dragonboard and ensured that
the test bailed out in the former and passed successfully on the latter.

Change-Id: I4b9cc3864f89dc665e767cf93b7a781db5bdd4c8
Reviewed-on: https://chromium-review.googlesource.com/307566
Commit-Ready: Roshan Pius <rpius@chromium.org>
Tested-by: Roshan Pius <rpius@chromium.org>
Reviewed-by: mukesh agrawal <quiche@chromium.org>