brillo_test_launcher: Add quickmerge.
Added a new feature to brillo_test_launcher called quickmerge.
This mimics test_that's quickmerge but uses a whitelist instead
of a blacklist. The whitelist is restricted to directories that
would contain test code and test libraries and will not impact
the currently running infrastructure.

This allows developers to quickly edit tests on their workstations
and run them on MobLab quickly, without the need for a Chrome OS
checkout and a chroot.

BUG=b:25122388
TEST=editted brillo_Gtests locally and verified the new version
     ran.

Change-Id: Ia974fac1d2f4f3fe9647fcffdf54f3fc89e1e266
Reviewed-on: https://chromium-review.googlesource.com/307862
Reviewed-by: Dan Shi <dshi@chromium.org>
Commit-Queue: Simran Basi <sbasi@chromium.org>
Tested-by: Simran Basi <sbasi@chromium.org>