[autotest] Implement preserve symlinks and perms for {get,send}_file on adbhost.
get_file:
If preserve_symlinks is True, then check if the source file is a symlink
and create the symlink on the drone.  If perms is True, chmod the same
perms on the source file once it's back on the drone.

send_file:
Create the symlink on the device if the source file is a symlink itself.

BUG=chromium:538793
TEST=Checked that get_file preserved symlinks and perms through modified
server dummy pass test.

Change-Id: I47b9bd31de7681549adfcdacc00668a0108395a0
Reviewed-on: https://chromium-review.googlesource.com/308197
Commit-Ready: Kevin Cheng <kevcheng@chromium.org>
Tested-by: Kevin Cheng <kevcheng@chromium.org>
Reviewed-by: Simran Basi <sbasi@chromium.org>