[autotest] Enforce a minimum version for deployment.
When running deployment_test (for a new deployment), or when running
repair_test (for manual repair), enforce a minimum version for a
board's stable repair version.

The enforced minimum version will be the _later_ of the current
stable repair image version and the currently preferred beta channel
image for the board.

BUG=None
TEST=Run with the new --nostable option to test the logic

Change-Id: I6a678172450f3923aaa52068cc6146e6867ecde4
Reviewed-on: https://chromium-review.googlesource.com/307864
Commit-Ready: Richard Barnette <jrbarnette@chromium.org>
Tested-by: Richard Barnette <jrbarnette@chromium.org>
Reviewed-by: Dan Shi <dshi@chromium.org>