graphics_GLBench: upload new MTK reference image
This image differs from the other (sensible) reference image in one
pixel in the lower left-hand corner, and the difference in question
is due solely to the alpha value - we set it to 127 whereas the other
image sets it to 255.

As far as we can tell, our image is correct.

BUG=chrome-os-partner:42626
TEST=ImageMagick compare to
     compositing_no_fill.pixmd5-642537874c2920ca7ddf48d08c17f74a.png

Change-Id: I2cc543beb7e7a27364fb626d241afee0ba236c1e
Signed-off-by: Rufus Hamade <rufus.hamade@imgtec.com>
Reviewed-on: https://chromium-review.googlesource.com/308780
Commit-Ready: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: cawa cheng <cawa.cheng@mediatek.com>
Reviewed-by: Ilja Friedel <ihf@chromium.org>