autotest(wifi): Move frequency check to wifi_test_context_manager.
Move the check to see if DUT supports the provided frequency
configuration to central |WifiTestContextManager|.configure so that
we don't need to add this check separately for every test.

BUG=25129221
TEST=Ran network_WiFi_SimpleConnect.wifi_check11a &
network_WiFi_RoamDbus on dragonboard and ensured
that the test bailed out in both the cases.

Change-Id: I1bbc4ac544df2a6dd30a27a8d5b66a7d3fd775b9
Reviewed-on: https://chromium-review.googlesource.com/308568
Commit-Ready: Roshan Pius <rpius@chromium.org>
Tested-by: Roshan Pius <rpius@chromium.org>
Reviewed-by: Roshan Pius <rpius@chromium.org>