autotest(wifi): Filter out unwanted beacons in pcap
network_WiFi_ChannelScanDwellTime expects all the beacons captured
during the test to be from the AP's configured with a specific prefix
during the test. This results in test failures if there are some
background AP's interfering. So, filter out the result before analyzing.

While there:
Add a check to ensure that we have some valid beacons in our capture
before procedding to analyze the captures.

BUG=b:25345601
TEST=`test_that --board=$board $host network_WiFi_ChannelScanDwellTime`
on Dragonboard.

Change-Id: I93690f65850b95842cead7ed79ca8218da70502a
Reviewed-on: https://chromium-review.googlesource.com/309620
Commit-Ready: Roshan Pius <rpius@chromium.org>
Tested-by: Roshan Pius <rpius@chromium.org>
Reviewed-by: mukesh agrawal <quiche@chromium.org>