power_LoadTest: Open trap window for auto loaded tabs
There's a tab that opens on first login for Chromebooks for free offers.
Since every login for power_LoadTest is a first login (since the state is
cleared), open a window to catch that tab (and any others) which will be
closed by the test later.

BUG=None
TEST=run power_LoadTest.1hour
Check that there are only 5 tabs in the background window

Change-Id: I650f636e23435c4f8e9a26b09a39ef378cd94068
Reviewed-on: https://chromium-review.googlesource.com/309543
Commit-Ready: Derek Basehore <dbasehore@chromium.org>
Tested-by: Derek Basehore <dbasehore@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>