bypass broken screen after attempting to bypass insert screen
This patch integrates broken screen bypass code with insert screen bypass
code.

It means we first attempt to bypass the insert screen and waits for a
ping connection. If the dut is running on old firmware, it should boot and
respond to ping. If it doesn't, it implies the dut is running on the new
firmware and stuck at the broken screen. Then, we take the broken screen
bypass action.

This is needed for the boards which have to run on both tot firmware and
branched firmware.

BUG=chromium:501060,chrome-os-partner:46752
BRANCH=tot
TEST=test_that -b veyron_jerry suite:faft_bios

Change-Id: I5bccb57ace1511f49522849c5ab8c1a3cf99d54d
Signed-off-by: Daisuke Nojiri <dnojiri@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/309558
Reviewed-by: Wai-Hong Tam <waihong@chromium.org>