[Autotest] Improvements to touch_UpdateErrors test
Make touch_UpdateErrors check that the correct device id checks for
update.  In this vein, redo the code in input_playback where the
hw_id is found, to fix the exceptions of sumo and lulu.
Now, the test will fail if a device with both a touchscreen and
touchpad skipped checking for updates on only one of them.

Also identify the boards with problematic FSI builds.  These boards,
such as expresso, have failures updating to their FSI firmware -
a problem we cannot fix.  Only check the most recent build's logs
for these devices.  Remove the hack that was there before.
All other devices behave the same way.
Now, the test will fail if, on one of the exception boards, there
is both a failure in the FSI build present and a failure on the
most recent build.

Modify touch_WakeupSource to split invalid boards into invalid
touchpads and invalid touchscreens.
Now, test will not fail on sumo.

BUG=None
TEST=ran affected tests on several devices

Change-Id: I697522f2c9992990a2d2adb0bb997a09fed5d4c7
Reviewed-on: https://chromium-review.googlesource.com/310312
Reviewed-by: David Haddock <dhaddock@chromium.org>
Tested-by: Katherine Threlkeld <kathrelkeld@google.com>
Tested-by: Katherine Threlkeld <kathrelkeld@chromium.org>
Commit-Queue: Katherine Threlkeld <kathrelkeld@chromium.org>