[autotest] Skip reboot to the default label on servo repair flow
The reboot method supports to reboot to a specific label (a label
here means a specific kernel). By default, it reboots to the latest
label. In Chrome OS world, we don't use this mechanism and should
skip it totally.

Otherwise, an error happens, complaining permission denied as the
tool for this reboot mechanism is placed to /tmp and Chrome OS
disables /tmp exec permission.

BUG=chromium:534882
TEST=Ran the following Python code for a DUT needs servo repair.
  import common
  from autotest_lib.server import hosts
  host = hosts.create_host(DUT_HOSTNAME, servo_args={})

Change-Id: Ie918062c3fe59f590e7f5b677148f3cac9f4031e
Reviewed-on: https://chromium-review.googlesource.com/310553
Commit-Ready: Wai-Hong Tam <waihong@chromium.org>
Tested-by: Wai-Hong Tam <waihong@chromium.org>
Reviewed-by: danny chan <dchan@chromium.org>
Reviewed-by: Dan Shi <dshi@chromium.org>