[autotest] Don't limit FAFT running on a subset of boards
The original way to limit FAFT to run on a subset of boards is
no longer scalable now. We remove this limitation. By default,
all DUTs with 'pool:faft-test' label are scheduled to run FAFT.

For experimenting new boards, we can first label the DUTs as
'pool:faft-test-experiment'. After we verify them work, we can
relabel them to 'pool:faft-test'.

BUG=chromium:534882
TEST=None

Change-Id: I98806bb921360c3976b5a09a2828cbbd5b3eb6bd
Reviewed-on: https://chromium-review.googlesource.com/311398
Reviewed-by: Dan Shi <dshi@chromium.org>
Tested-by: Wai-Hong Tam <waihong@chromium.org>