Fix handling in assign_stable_images when default is unchanged.
In assign_stable_images, if a new board should be updated to the
default build, but the default build was unchanged, then the update
wouldn't take place.  This fixes it to make all changes to convert
a board to default to happen unconditionally, rather than only when
the default board is changing.

Additionally, this improves message output to describe the cases of
interest.

BUG=None
TEST=use a dry run

Change-Id: Idbdc0fb10c37c29c9749c3c7c454bd6f6d88e735
Reviewed-on: https://chromium-review.googlesource.com/312280
Commit-Ready: Richard Barnette <jrbarnette@chromium.org>
Tested-by: Richard Barnette <jrbarnette@chromium.org>
Reviewed-by: Dan Shi <dshi@chromium.org>