[autotest] In deployment_test, put stdout into logs, too.
The `deployment_test` and `repair_test` commands write a short
report summary to stdout.  This report is lost unless the user
copies it into a file for later.  Having the report is convenient
tool for debugging failures, so this changes the script to save
the report in the log directory under the name 'report.log'.

BUG=None
TEST=run with --noinstall --nostable

Change-Id: Iea65fc6a3daaf452a79da3c9b4d743fd224cbc55
Reviewed-on: https://chromium-review.googlesource.com/312741
Tested-by: Richard Barnette <jrbarnette@chromium.org>
Reviewed-by: Kevin Cheng <kevcheng@chromium.org>