Implement policy_URLBlacklist AutoTest
This test verifies the behavior of Chrome OS for a range of valid
values of the URLBlacklist user policy, using three unique
test cases: NotSet, SingleBlacklistedFile, MultipleBlacklistedFiles.

When the policy is undefined (case=NotSet), None of the URLs are blocked.
When set to one or more URLs, navigation to the URLs
with the same domain as any of the listed URL/domains is blocked

BUG=chromium:556692
TEST=Run test_that <IP> policy_URLBlacklist

Change-Id: I693e5602438e27d513e37fef42f5c88ef24e086d
Reviewed-on: https://chromium-review.googlesource.com/312891
Commit-Ready: Krishna Gavini <krishnargv@chromium.org>
Tested-by: Krishna Gavini <krishnargv@chromium.org>
Reviewed-by: Jacob Dufault <jdufault@chromium.org>