[autotest] Add in testbed class and creation method in host factory.
The testbed class for now will assume all devices are accessible via ADB
and that there is only 1 test station.  It will not create the ADBHost
like in factory and mix in a bunch of other host types with the host
(mostly because it's unnecessary).

The current abilities of the testbed class is just to retrieve the host
types within it.
- get_all_hosts
- get_test_station
- get_adb_devices

BUG=chromium:548739
TEST=With a moblab and 3 adb devices (shamu, hammerhead, angler),
ensured successful creation of testbed object as well as successful run of
ADBHost methods for each adb devices (get_board and run).

Change-Id: I9045f94050ad29421b64924679891a7a5906e294
Reviewed-on: https://chromium-review.googlesource.com/313223
Commit-Ready: Kevin Cheng <kevcheng@chromium.org>
Tested-by: Kevin Cheng <kevcheng@chromium.org>
Reviewed-by: Kevin Cheng <kevcheng@chromium.org>