graphics_Idle: Simplify GPU detection and make it more scalable
The board detection is a little limited in that it doesn't
automagically scale as we add more boards. Instead of having a list of
GPUs in each test, we add a generic test for intel, and then filter
out the GPUs where the test isn't supposed to run.

BUG=chromium:557972
TEST=not tested

Change-Id: I59d368815ae7b82629c2fa90d48ff850046beb2e
Reviewed-on: https://chromium-review.googlesource.com/313445
Commit-Ready: Ilja Friedel <ihf@chromium.org>
Tested-by: Ilja Friedel <ihf@chromium.org>
Reviewed-by: Ilja Friedel <ihf@chromium.org>