[autotest] Clean up servo and chameleon hosts in CrosHost.
This adds a call in CrosHost.close() to close the servo host and
chameleon host, if any.  Without this call, ssh master connections
may be left behind indefinitely running on the drone or shard.

This symptom could show up in many cases, but was particularly easy
to see with the `repair_test` and `deployment_test` commands.

BUG=None
TEST=run `repair_test`, observe the servo master connection is cleaned up.

Change-Id: Iac611c6150985ed703b56ebcce2de343c33b29d0
Reviewed-on: https://chromium-review.googlesource.com/313661
Commit-Ready: Richard Barnette <jrbarnette@chromium.org>
Tested-by: Richard Barnette <jrbarnette@chromium.org>
Reviewed-by: Dan Shi <dshi@chromium.org>
Reviewed-by: Kevin Cheng <kevcheng@chromium.org>