autotest(wifi): Copy Android RPC server to AdbHost
Copy over a copy of the Android XML RPC server daemon to the AdbHost
machine before starting a test. This helps to ensure that the Android
XML RPC server version is in sync with the Autotest server.
This might be a temporary solution until we can finalize on a proper
deployment mechanism for this script on non-Cros Adbhost machines.

BUG=b/25818227
TEST=`test_that --board=panther android1758-wifi-test-station-2.cros
network_WiFi_SimpleConnect.wifi_check11b --args use_wpa_cli=0`

Change-Id: I3d610faf2afa2b5c3b56076028262153b05f0aea
Reviewed-on: https://chromium-review.googlesource.com/313646
Commit-Ready: Roshan Pius <rpius@chromium.org>
Tested-by: Roshan Pius <rpius@chromium.org>
Reviewed-by: Roshan Pius <rpius@chromium.org>
Reviewed-by: Simran Basi <sbasi@chromium.org>