[autotest] Make `atest stable_version list` readable again.
The output of the `atest stable_version list` command used
fixed-width columns in its output, and those widths were much to
small for typical output.  The result was that the output was
not tabular, and largely not readable.

This changes the command to use dynamically selected column
widths, so that the output will always be tabular.

BUG=None
TEST=run the command, see reasonable output

Change-Id: Id1ee3a74400caf9b3258d7cba2b91d9cae4fe895
Reviewed-on: https://chromium-review.googlesource.com/314289
Commit-Ready: Richard Barnette <jrbarnette@chromium.org>
Tested-by: Richard Barnette <jrbarnette@chromium.org>
Reviewed-by: Dan Shi <dshi@chromium.org>